+++ vllm/spec_decode/interfaces.py
